requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
requests==2.32.3
jinja2==3.1.6
python-multipart==0.0.20
pystray==0.19.5
numpy==2.4.4
Pillow==12.2.0
